requirements.txt
# Core dependencies
streamlit>=1.45.0
google-generativeai>=0.8.0
python-dotenv>=0.19.0
SpeechRecognition>=3.14.0
PyAudio>=0.2.14

# Data handling and visualization
pandas>=2.0.0
plotly>=6.0.0
numpy>=1.24.0

# Development dependencies
black  # Code formatting
pylint  # Code linting
pytest  # Testing

# Logging (built into Python, but explicit for clarity)
# logging is part of Python standard library
